+++ internal/tester/scenarios/s1enb/connectivity_dualstack.go
@@ -67,7 +67,7 @@ func runS1ENBConnectivityDualStack(ctx context.Context, env scenarios.Env, _ any
 	defer func() { _ = e.Close() }()
 
 	ue := e.NewUE(connDualStackIMSI, k, opc)
-	ue.RequestPDNType(eps.PDNTypeIPv4v6)
+	ue.RequestPDNType(uint8(eps.PDNTypeIPv4v6))
 
 	res, err := e.Attach(ue, 15*time.Second)
 	if err != nil {
