+++ internal/tester/scenarios/s1enb/static_ip.go
@@ -91,7 +91,7 @@ func runS1ENBStaticIP(ctx context.Context, env scenarios.Env, p *staticIPParams,
 
 	ue := e.NewUE(imsi, k, opc)
 	if ipv6 {
-		ue.RequestPDNType(eps.PDNTypeIPv6)
+		ue.RequestPDNType(uint8(eps.PDNTypeIPv6))
 	}
 
 	res, err := e.Attach(ue, 15*time.Second)
